+++ src/modwire_architecture/__init__.py
@@ -1,11 +1,27 @@
 from ._version import __version__, version
-from .architecture import ArchitectureConfig
+from .architecture import (
+    AnalysisEvidence,
+    AnalysisOutcome,
+    AnalysisStatus,
+    ArchitectureAnalysis,
+    ArchitectureConfig,
+    ArchitectureInsight,
+    EvidenceFact,
+    FindingSeverity,
+)
 from .facade import Modwire
 from .shared import ModwireConfigModel, ModwireModel, ModwireReportModel
 
 __all__ = [
     "Modwire",
     "ArchitectureConfig",
+    "ArchitectureAnalysis",
+    "ArchitectureInsight",
+    "AnalysisEvidence",
+    "AnalysisOutcome",
+    "AnalysisStatus",
+    "EvidenceFact",
+    "FindingSeverity",
     "ModwireConfigModel",
     "ModwireModel",
     "ModwireReportModel",